+++ trixnity-client/src/commonMain/kotlin/de/connect2x/trixnity/client/store/cache/FullRepositoryObservableCache.kt
@@ -22,23 +22,28 @@ private class FullRepositoryObservableCacheIndex<K>(private val loadFromStore: s
     private val subscribers = MutableStateFlow(0)
     private val fullyLoadedFromRepository = MutableStateFlow(false)
 
+    context(transaction: CacheTransaction)
     override suspend fun onPut(key: K) {
         allKeys.add(key)
     }
 
+    context(transaction: CacheTransaction)
     override suspend fun onSkipPut(key: K) {
         fullyLoadedFromRepository.value = false
     }
 
+    context(transaction: CacheTransaction)
     override suspend fun onRemove(key: K, stale: Boolean) {
         fullyLoadedFromRepository.update { it && stale }
         allKeys.remove(key)
     }
 
+    context(transaction: CacheTransaction)
     override suspend fun onRemoveAll() {
         allKeys.removeAll()
     }
 
+    context(transaction: CacheTransaction)
     override suspend fun getSubscriptionCount(key: K): Int = subscribers.value
 
     fun getAllKeys(): Flow<Set<K>> =
@@ -95,5 +100,5 @@ internal open class FullRepositoryObservableCache<K : Any, V>(
         addIndex(subscribersIndex)
     }
 
-    fun readAll(): Flow<Map<K, Flow<V?>>> = subscribersIndex.getAllKeys().map { keys -> keys.associateWith { get(it) } }
+    fun getAll(): Flow<Map<K, Flow<V?>>> = subscribersIndex.getAllKeys().map { keys -> keys.associateWith { get(it) } }
 }