+++ trixnity-client/src/commonMain/kotlin/de/connect2x/trixnity/client/store/cache/MapRepositoryObservableCache.kt
@@ -32,11 +32,13 @@ private class MapRepositoryObservableCacheIndex<K1 : Any, K2>(
 
     private val values = ConcurrentObservableMap<K1, MapRepositoryObservableMapIndexValue<K2>>()
 
+    context(transaction: CacheTransaction)
     override suspend fun onPut(key: MapRepositoryCoroutinesCacheKey<K1, K2>) {
         log.trace { "$name: put key $key into map index" }
         values.getOrPut(key.firstKey) { MapRepositoryObservableMapIndexValue() }.keys.add(key.secondKey)
     }
 
+    context(transaction: CacheTransaction)
     override suspend fun onSkipPut(key: MapRepositoryCoroutinesCacheKey<K1, K2>) {
         values.update(key.firstKey) { mapping ->
             if (mapping != null && mapping.fullyLoadedFromStore) {
@@ -46,6 +48,7 @@ private class MapRepositoryObservableCacheIndex<K1 : Any, K2>(
         }
     }
 
+    context(transaction: CacheTransaction)
     override suspend fun onRemove(key: MapRepositoryCoroutinesCacheKey<K1, K2>, stale: Boolean) {
         values.update(key.firstKey) { mapping ->
             if (mapping != null) {
@@ -61,30 +64,32 @@ private class MapRepositoryObservableCacheIndex<K1 : Any, K2>(
         }
     }
 
+    context(transaction: CacheTransaction)
     override suspend fun onRemoveAll() {
         values.removeAll()
     }
 
+    context(transaction: CacheTransaction)
     override suspend fun getSubscriptionCount(key: MapRepositoryCoroutinesCacheKey<K1, K2>): Int =
         values.getOrPut(key.firstKey) { MapRepositoryObservableMapIndexValue() }.subscribers.value
 
-    fun getMapping(key: K1): Flow<Set<K2>> = flow {
-        val value = values.update(key) { it ?: MapRepositoryObservableMapIndexValue(fullyLoadedFromStore = false) }
-        checkNotNull(value)
-        emitAll(
-            flow {
-                    val fullyLoadedFromStore = values.get(key)?.fullyLoadedFromStore
-                    if (fullyLoadedFromStore != true) {
-                        log.trace { "$name: not fully loaded from store. load now for key $key" }
-                        loadFromStore(key)
-                    }
-                    values.update(key) { it?.copy(fullyLoadedFromStore = true) }
-                    emitAll(value.keys.values)
+    fun getMapping(key: K1): Flow<Set<K2>> =
+        flow {
+                val value = withCacheTransaction { values.getOrPut(key) { MapRepositoryObservableMapIndexValue() } }
+                val fullyLoadedFromStore = value.fullyLoadedFromStore
+                if (!fullyLoadedFromStore) {
+                    log.trace { "$name: not fully loaded from store. load now for key $key" }
+                    loadFromStore(key)
+                    withCacheTransaction { values.update(key) { it?.copy(fullyLoadedFromStore = true) } }
                 }
-                .onStart { value.subscribers.update { it + 1 } }
-                .onCompletion { value.subscribers.update { it - 1 } }
-        )
-    }
+                emitAll(value.keys.values)
+            }
+            .onStart {
+                withCacheTransaction { values.getOrPut(key) { MapRepositoryObservableMapIndexValue() } }
+                    .subscribers
+                    .update { it + 1 }
+            }
+            .onCompletion { values.get(key)?.subscribers?.update { it - 1 } }
 
     override suspend fun collectStatistic(): ObservableCacheIndexStatistic {
         val (all, subscribed) = values.internalRead { count() to values.count { it.subscribers.value > 0 } }
@@ -125,7 +130,7 @@ internal open class MapRepositoryObservableCache<K1 : Any, K2, V>(
         addIndex(mapRepositoryIndex)
     }
 
-    fun readByFirstKey(key: K1): Flow<Map<K2, Flow<V?>>> =
+    fun getByFirstKey(key: K1): Flow<Map<K2, Flow<V?>>> =
         mapRepositoryIndex.getMapping(key).map { mapping ->
             mapping.associateWith { secondKey -> get(MapRepositoryCoroutinesCacheKey(key, secondKey)) }
         }